Security: JWT authentication and User entity with its layers (#9)
* Security: JWT authentication and User entity with its layers

* refactor: modify authorization config and small fixes

* add missed dependencies

* refactor: bugfix - identifiers type

* reset settings in app.properties

+++ src/main/java/com/example/carsharing/service/RentalService.java
@@ -5,11 +5,11 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface RentalService {
-    Rental create(LocalDate returnDate, Integer carId, Integer userId);
+    Rental create(LocalDate returnDate, Long carId, Long userId);
 
-    List<Rental> getByUserIdAndStatus(Integer id, boolean isActive);
+    List<Rental> getByUserIdAndStatus(Long id, boolean isActive);
 
-    Rental getById(Integer id);
+    Rental getById(Long id);
 
     Rental setActualReturnDate(LocalDate actualReturnDate, Rental rental);
 }
